changed text color of the buttton when its disabled

+++ bandits-api/src/main/java/com/bandits/api/bandits_api/repositories/RecipeRepository.java
@@ -8,9 +8,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+public interface RecipeRepository extends JpaRepository<Recipe, Integer> {
 
     Recipe findByUri(String uri);
+
     List<Recipe> findByUser(User user);
 
 }
